repos/repo-postmerge/gentoo: change from 6h->1h for news
I'm not sure why this delay exists at all -- why not pull every time
we do ::gentoo, as we really want them in lockstep? We *very often*
push news items in sync with ebuild changes.

Anyway, I'm not 100% sure if this is right, given that it feels odd
to ever want a 6h delay, and also that it hasn't come up before. But
let's try it.

Noticed because the wireplumber news item appeared via rsync but
not via the gentoo-mirror / sync (w/ metadata) git repos. Let's try changing
6h -> 1h to see if it helps...

Signed-off-by: Sam James <sam@gentoo.org>